requirements.txt
pillow>=9.5.0
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.1
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
seaborn==0.12.2
scipy==1.11.2
imutils==0.5.4
scikit-image==0.21.0 